Add EOTG Page Setup
