--- src/pages/EOTGPage/EOTGLayout.ts
@@ -1,15 +1,0 @@
-import React from "react";
-
-class EOTGLayout extends React.Component {
-  render() {
-    return (
-      <>
-        <div id="eotg-main">
-          {React.cloneElement(this.props.children, { ...this.props })}
-        </div>
-      </>
-    );
-  }
-}
-
-export default EOTGLayout;
