+++ src/pages/EOTGPage/components/About.css
@@ -1,0 +1,71 @@
+#eotg-main {
+  height: 100%;
+  background: #71c167; /* $eotg-green */
+}
+
+.eotg-contact {
+  background: #71c167; /* $eotg-green */
+}
+
+.eotg-contact a {
+  color: white;
+}
+
+/* About Component CSS */
+.eotg-about {
+  /* background-image: url(../../../../pages/EOTGPage/svg/bg-about-winter.svg); */
+  /* Uncomment and replace the above URL to change the seasons */
+  background-size: contain;
+  background-color: #84dadf; /* $eotg-blue */
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  position: relative;
+}
+
+.eotg-about h1 {
+  padding-top: 5rem;
+  font-size: 3rem;
+  font-family: "Ubuntu", sans-serif;
+  color: #e8f7fa;
+}
+
+.eotg-about p,
+.eotg-about a {
+  font-size: 2rem;
+  font-family: "Ubuntu", sans-serif;
+  color: #e8f7fa;
+}
+
+.eotg-about a {
+  text-decoration: underline;
+}
+
+/* Responsive design (media query for small screens) */
+@media (max-width: 576px) {
+  .eotg-about {
+    padding-top: 1rem;
+  }
+}
+
+/* Floating animation */
+.floating {
+  animation-name: floating;
+  animation-duration: 3s;
+  animation-iteration-count: infinite;
+  animation-timing-function: ease-in-out;
+  margin-left: 30px;
+  margin-top: 5px;
+}
+
+@keyframes floating {
+  0% {
+    transform: translate(0, 0px);
+  }
+  50% {
+    transform: translate(0, 15px);
+  }
+  100% {
+    transform: translate(0, -0px);
+  }
+}
