+++ src/pages/EOTGPage/components/Hero.css
@@ -1,0 +1,93 @@
+.eotg-hero {
+  background-image: url(../../../../pages/UpcomingEvents/EOTGPage/svg/bg-sunset.svg);
+  background-size: cover;
+  min-height: 100vh;
+  display: flex;
+  position: relative;
+}
+
+.eotg-hero__container {
+  width: 100%;
+  height: 100%;
+}
+
+.eotg-hero__row {
+  height: 100%;
+}
+
+.eotg-hero__title {
+  font-size: 3rem;
+  font-family: "Ubuntu", sans-serif;
+  color: #264653;
+  animation-name: floating;
+  animation-duration: 3s;
+  animation-iteration-count: infinite;
+  animation-timing-function: ease-in-out;
+  margin-left: 30px;
+  margin-top: 5px;
+}
+
+.eotg-hero__graphic {
+  width: 100%;
+  align-self: flex-end;
+}
+
+.eotg-hero__logo {
+  width: 60%;
+}
+
+.eotg-hero__detail {
+  font-size: 1.5rem;
+  font-family: "Ubuntu", sans-serif;
+  color: #264653;
+}
+
+.eotg-hero__button_div {
+  margin-top: 2rem;
+}
+
+.eotg-hero__button {
+  border: 4px solid white;
+  border-radius: 16px;
+  color: white;
+  background: transparent;
+  padding-left: 0.5rem;
+  padding-right: 0.5rem;
+  padding-top: 0.75rem;
+  padding-bottom: 0.75rem;
+  margin-top: 1rem;
+  font-size: 1.2rem;
+}
+
+.eotg-hero__button:hover {
+  background: white;
+  color: #71c167; /* $eotg-green */
+  text-decoration: none;
+}
+
+.eotg-hero .bg-img {
+  -webkit-user-drag: none;
+  user-select: none;
+}
+
+/* Floating animation */
+.floating {
+  animation-name: floating;
+  animation-duration: 3s;
+  animation-iteration-count: infinite;
+  animation-timing-function: ease-in-out;
+  margin-left: 30px;
+  margin-bottom: 5rem;
+}
+
+@keyframes floating {
+  0% {
+    transform: translate(0, 0px);
+  }
+  50% {
+    transform: translate(0, 15px);
+  }
+  100% {
+    transform: translate(0, -0px);
+  }
+}
