+++ src/pages/EOTGPage/data/OrgGalleries.ts
